move model helper function in pipeline to EfficiencyMixin
