+++ src/diffusers/__init__.py
@@ -121,6 +121,7 @@ else:
             "DDPMPipeline",
             "DiffusionPipeline",
             "DiTPipeline",
+            "EfficiencyMixin",
             "ImagePipelineOutput",
             "KarrasVePipeline",
             "LDMPipeline",
@@ -505,6 +506,7 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
             DDPMPipeline,
             DiffusionPipeline,
             DiTPipeline,
+            EfficiencyMixin,
             ImagePipelineOutput,
             KarrasVePipeline,
             LDMPipeline,
