+++ src/diffusers/pipelines/__init__.py
@@ -48,6 +48,7 @@ else:
     _import_structure["pipeline_utils"] = [
         "AudioPipelineOutput",
         "DiffusionPipeline",
+        "EfficiencyMixin",
         "ImagePipelineOutput",
     ]
     _import_structure["deprecated"].extend(
@@ -328,6 +329,7 @@ if TYPE_CHECKING or DIFFUSERS_SLOW_IMPORT:
         from .pipeline_utils import (
             AudioPipelineOutput,
             DiffusionPipeline,
+            EfficiencyMixin,
             ImagePipelineOutput,
         )
